refactor: move yield rewards logic to suite common
