+++ suite-common/earn-stablecoin-api/package.json
@@ -24,8 +24,13 @@
         "@suite-common/http-client": "workspace:^",
         "@suite-common/react-query": "workspace:^",
         "@suite-common/schemas": "workspace:^",
+        "@suite-common/wallet-config": "workspace:^",
+        "@suite-common/wallet-types": "workspace:^",
+        "@trezor/blockchain-link-types": "workspace:^",
         "@trezor/env-utils": "workspace:*",
+        "@trezor/react-utils": "workspace:^",
         "@trezor/utils": "workspace:^",
+        "react": "19.2.4",
         "zod": "4.3.6"
     }
 }
