+++ suite-common/earn-stablecoin-api/src/index.ts
@@ -1,5 +1,3 @@
-export * from './api/schemas';
-export * from './api/types';
 export * from './services';
 export * from './config';
 export * from './verification';
@@ -7,6 +5,6 @@ export * from './hooks/useAllYieldOpportunities';
 export * from './hooks/useGetYieldOpportunities';
 export * from './hooks/useEnterYieldOpportunity';
 export * from './hooks/useExitYieldOpportunity';
-export * from './hooks/useGetMerkleRewards';
+export * from './hooks/merkle-rewards';
 export * from './constants';
 export * from './utils/sortRewardsByUnderlyingToken';
