+++ suite-common/earn-stablecoin-api/tsconfig.json
@@ -7,6 +7,7 @@
         { "path": "../react-query" },
         { "path": "../schemas" },
         { "path": "../../packages/env-utils" },
+        { "path": "../../packages/react-utils" },
         { "path": "../../packages/utils" }
     ]
 }
