+++ suite-common/wallet-config/src/types.ts
@@ -1,4 +1,4 @@
-import { type NetworkDtoId } from '@suite-common/earn-stablecoin-api';
+import { type NetworkDtoId } from '@suite-common/earn-stablecoin-api/src/api';
 import { type DeviceModelInternal } from '@trezor/device-utils';
 import { type RequiredKey } from '@trezor/type-utils';
